MockUser32をEditorTestSuiteに移動する
